+++ src/scripts/components/layer-selector/layer-selector.tsx
@@ -1,6 +1,9 @@
 import React, {FunctionComponent, useEffect, useState} from 'react';
 import {useSelector, useDispatch} from 'react-redux';
+import {useIntl} from 'react-intl';
+
 import {layersSelector} from '../../reducers/layers';
+import {selectedLayersSelector} from '../../reducers/selected-layers';
 import fetchLayers from '../../actions/fetch-layers';
 import {setSelectedLayerIdAction} from '../../actions/set-selected-layer';
 import LayerList from '../layer-list/layer-list';
@@ -8,21 +11,43 @@ import Tabs from '../tabs/tabs';
 import styles from './layer-selector.styl';
 
 const LayerSelector: FunctionComponent<{}> = () => {
+  const intl = useIntl();
   const layers = useSelector(layersSelector);
+  const layerIds = useSelector(selectedLayersSelector);
   const dispatch = useDispatch();
   const tabs = [
     {
       id: 'main',
-      label: 'Main'
+      label: intl.formatMessage({id: 'layerSelector.main'})
     },
     {
       id: 'compare',
-      label: 'Compare'
+      label: intl.formatMessage({id: 'layerSelector.compare'})
     }
   ];
 
   const [activeTabId, setActiveTabId] = useState(tabs[0].id);
+  const [isOpen, setIsOpen] = useState(false);
   const isMainTabSelected = activeTabId === tabs[0].id;
+  const selectedLayer = isMainTabSelected ? layerIds.main : layerIds.compare;
+
+  const onTabClick = (id: string) => {
+    setActiveTabId(id);
+
+    if (!isOpen) {
+      setIsOpen(true);
+      return;
+    }
+
+    if (activeTabId === id) {
+      setIsOpen(false);
+    }
+  };
+
+  const onLayerClick = (id: string) => {
+    const newId = selectedLayer === id ? null : id;
+    dispatch(setSelectedLayerIdAction(newId, isMainTabSelected));
+  };
 
   useEffect(() => {
     dispatch(fetchLayers());
@@ -33,15 +58,15 @@ const LayerSelector: FunctionComponent<{}> = () => {
       <Tabs
         tabs={tabs}
         activeTabId={activeTabId}
-        onTabChanged={id => setActiveTabId(id)}
+        onTabChanged={id => onTabClick(id)}
       />
-      <LayerList
-        layers={layers}
-        selected={'layer1'}
-        onSelect={id =>
-          dispatch(setSelectedLayerIdAction(id, isMainTabSelected))
-        }
-      />
+      {isOpen && (
+        <LayerList
+          layers={layers}
+          selected={selectedLayer}
+          onSelect={id => onLayerClick(id)}
+        />
+      )}
     </div>
   );
 };