+++ src/scripts/components/tab/tab.tsx
@@ -1,6 +1,6 @@
 import React, {FunctionComponent} from 'react';
+import cx from 'classnames';
 import styles from './tab.styl';
-import classNames from 'classnames';
 
 interface Props {
   id: string;
@@ -14,11 +14,11 @@ const Tab: FunctionComponent<Props> = ({
   onSelectTabId,
   children
 }) => {
-  const btnClass = classNames(styles.btn, {
-    [styles.btnInactive]: activeTabId !== id
-  });
+  const isActive = activeTabId === id;
+  const tabClasses = cx(styles.tab, isActive && styles.tabActive);
+
   return (
-    <button className={btnClass} onClick={() => onSelectTabId(id)}>
+    <button className={tabClasses} onClick={() => onSelectTabId(id)}>
       {children}
     </button>
   );