+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+# 0.2.0 (2019-10-11)
+
+
+### Bug Fixes
+
+* **globe:** hide bottom container ([#144](https://github.com/ubilabs/esa-climate-from-space/issues/144)) ([5ee1cab](https://github.com/ubilabs/esa-climate-from-space/commit/5ee1cab))
+* **webpack:** remove source maps from production build ([#3](https://github.com/ubilabs/esa-climate-from-space/issues/3)) ([76984c5](https://github.com/ubilabs/esa-climate-from-space/commit/76984c5))
+
+
+### Features
+
+* **cesium:** add first cesium globe ([#80](https://github.com/ubilabs/esa-climate-from-space/issues/80)) ([5f91047](https://github.com/ubilabs/esa-climate-from-space/commit/5f91047))
+* **cesium:** use static image provider ([#140](https://github.com/ubilabs/esa-climate-from-space/issues/140)) ([b0c2d07](https://github.com/ubilabs/esa-climate-from-space/commit/b0c2d07))
+* **ci:** add cloud build config ([d5eb64d](https://github.com/ubilabs/esa-climate-from-space/commit/d5eb64d))
+* **eslint:** add eslint ([#5](https://github.com/ubilabs/esa-climate-from-space/issues/5)) ([01882e8](https://github.com/ubilabs/esa-climate-from-space/commit/01882e8))
+* **globe:** add second globe ([#133](https://github.com/ubilabs/esa-climate-from-space/issues/133)) ([2d4a0f4](https://github.com/ubilabs/esa-climate-from-space/commit/2d4a0f4))
+* **globe:** sync globes ([6f33f5b](https://github.com/ubilabs/esa-climate-from-space/commit/6f33f5b))
+* **i18n:** add language switcher ([44610cb](https://github.com/ubilabs/esa-climate-from-space/commit/44610cb))
+* **i18n:** add react-intl and local state ([79766f0](https://github.com/ubilabs/esa-climate-from-space/commit/79766f0))
+* **i18n:** add react-intl and locale state ([#130](https://github.com/ubilabs/esa-climate-from-space/issues/130)) ([549f3eb](https://github.com/ubilabs/esa-climate-from-space/commit/549f3eb)), closes [#131](https://github.com/ubilabs/esa-climate-from-space/issues/131)
+* **i18n:** detect browser language ([d41d200](https://github.com/ubilabs/esa-climate-from-space/commit/d41d200))
+* **i18n:** refetch layers when locale changes ([60ce6c0](https://github.com/ubilabs/esa-climate-from-space/commit/60ce6c0))
+* **layer-component:** add styles ([#120](https://github.com/ubilabs/esa-climate-from-space/issues/120)) ([ea71095](https://github.com/ubilabs/esa-climate-from-space/commit/ea71095))
+* **layer-tabs:** add collapsible layer-selector, highlight selected … ([#123](https://github.com/ubilabs/esa-climate-from-space/issues/123)) ([2ed8363](https://github.com/ubilabs/esa-climate-from-space/commit/2ed8363))
+* **layers:** fetch layer definitions ([#79](https://github.com/ubilabs/esa-climate-from-space/issues/79)) ([13a7199](https://github.com/ubilabs/esa-climate-from-space/commit/13a7199))
+* **menu:** add menu-button with menu-items ([#124](https://github.com/ubilabs/esa-climate-from-space/issues/124)) ([753e948](https://github.com/ubilabs/esa-climate-from-space/commit/753e948))
+* **setup:** add build info ([#4](https://github.com/ubilabs/esa-climate-from-space/issues/4)) ([611a37a](https://github.com/ubilabs/esa-climate-from-space/commit/611a37a))
+* **setup:** setup project ([#2](https://github.com/ubilabs/esa-climate-from-space/issues/2)) ([63db92f](https://github.com/ubilabs/esa-climate-from-space/commit/63db92f))
+* **stories:** add story items and styles ([#142](https://github.com/ubilabs/esa-climate-from-space/issues/142)) ([52af98e](https://github.com/ubilabs/esa-climate-from-space/commit/52af98e))
+* **story-selector:** add story selector style ([#143](https://github.com/ubilabs/esa-climate-from-space/issues/143)) ([9c1b089](https://github.com/ubilabs/esa-climate-from-space/commit/9c1b089))
+
+
+
 ## 0.1.1 (2019-10-01)